Added refactored base types
