+++ src/types/response/base/Tweet.ts
